requirements.txt
langchain 
langgraph 
google-generativeai 
google-ai-generativelanguage 
yfinance
dotenv
langchain-community
langchain-google-genai
langgraph_supervisor
faiss-cpu
pypdf
streamlit
SpeechRecognition
gtts
pydub
fastapi
uvicorn
python-multipart
audio_recorder_streamlit
